+++ packages/grid/x-grid/src/tests/selection.XGrid.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { expect } from 'chai';
 import { spy } from 'sinon';
-import { getColumnValues } from 'test/utils/helperFn';
+import { getColumnValues, getRow } from 'test/utils/helperFn';
 import {
   // @ts-expect-error need to migrate helpers to TypeScript
   screen,
@@ -161,5 +161,34 @@ describe('<XGrid /> - Selection', () => {
     expect(getSelectedRows(apiRef)).to.deep.equal([2]);
     fireEvent.click(selectAll);
     expect(getSelectedRows(apiRef)).to.deep.equal([]);
+  });
+
+  it('should only select visible rows on the current page', () => {
+    render(
+      <div style={{ width: 300, height: 300 }}>
+        <XGrid
+          rows={[
+            {
+              id: 0,
+              brand: 'Nike',
+            },
+            {
+              id: 1,
+              brand: 'Puma',
+            },
+          ]}
+          columns={[{ field: 'brand', width: 100 }]}
+          checkboxSelection
+          checkboxSelectionVisibleOnly
+          pagination
+          pageSize={1}
+        />
+      </div>,
+    );
+    const selectAllCheckbox = document.querySelector('input[type="checkbox"]');
+    fireEvent.click(selectAllCheckbox);
+    expect(getRow(0)).to.have.class('Mui-selected');
+    fireEvent.click(screen.getByRole('button', { name: /next page/i }));
+    expect(getRow(1)).not.to.have.class('Mui-selected');
   });
 });
